refactor: moverContext파일 정리 및 index파일 구조 제거
