+++ reducers/moverReducer.tsx
