+++ src/components/domain/detail/DetailSections.tsx
