+++ src/components/domain/detail/LineDivider.tsx
