+++ src/components/domain/detail/MainMoverCard.tsx
