+++ src/components/domain/detail/ShareSection.tsx
