+++ src/components/domain/mover-search/FilterAreaServiceBox.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useMover } from "@/context/mover";
+import { useMover } from "@/lib/hooks/useMover";
 import Dropdown from "./Dropdown";
 import { DropdownOption } from "@/lib/types/mover.types"
