+++ src/components/domain/mover-search/FindDriverLayout.tsx
@@ -6,7 +6,9 @@ import FilterAreaServiceBox from "./FilterAreaServiceBox";
 import SearchBar from "./SearchBar";
 import { DropdownOption } from "@/lib/types/mover.types"
 import FavoriteDriverList from "./FavoriteDriverList";
-import { MoverProvider, useMover } from "@/context/mover";
+import { useMover } from "@/lib/hooks/useMover";
+import { MoverProvider } from "@/context/MoverContext"
+
 import {
   AREA_OPTIONS,
   SERVICE_OPTIONS,
