--- src/context/mover/index.tsx
@@ -1,3 +1,0 @@
-// index.ts
-export * from "./MoverContext";
-export * from "./useMover";
