--- src/lib/api/mover/index.ts
@@ -1,5 +1,0 @@
-// index.ts
-export * from "./getMover";
-export * from "./favoriteMover";
-export * from "@/lib/types/mover.types";
-export * from "@/constants/mover.constants";
